Refactor packages name and location
