+++ src/main/java/com/ahming/ui/DefaultFrame.java
@@ -1,14 +1,14 @@
-package com.ahming.components;
+package com.ahming.ui;
+
 
 import javax.swing.*;
 
 public class DefaultFrame extends JFrame {
-
     public DefaultFrame() {
         this.setTitle("AhMing's Keyboard Script Reader");
         this.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
         this.setAlwaysOnTop(true);
-        this.setResizable(true);
+        this.setResizable(false);
         this.setVisible(true);
 
         this.setSize(530, 300);
